ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
httpx==0.27.2

# RAG
langchain==0.3.1
langchain-community==0.3.1
langchain-groq==0.2.1
langchain-text-splitters==0.3.0
chromadb==0.5.11
sentence-transformers==3.1.1

# Document parsing
PyPDF2==3.0.1
python-pptx==1.0.2
python-docx==1.1.2

# Utilities
python-multipart==0.0.12
